BiocVersion-pkg {BiocVersion}	R Documentation

Determine Bioconductor version in use

Description

The 'major' and 'minor' version of this package determines the version of Bioconductor software packages to be installed by the install function.

Examples


packageVersion("BiocVersion")[,1:2]

[Package BiocVersion version 3.8.0 Index]